モデルの表示完了　しかし、モデルが高速で点滅する

+++ FileManager.cpp
@@ -68,7 +68,7 @@ Model* FileManager::loadModel(OBJECT obj)
 
             if (uniqueVertices.count(vertex) == 0)
             {
-                uniqueVertices[vertex] = static_cast<uint32_t>(model->getVerticesSize());
+                uniqueVertices[vertex] = model->getVerticesSize();
                 model->pushBackVertex(&vertex);
             }
